+++ src/main/java/org/gbif/occurrence/annotation/controller/RuleController.java
@@ -13,7 +13,6 @@
  */
 package org.gbif.occurrence.annotation.controller;
 
-import org.gbif.api.vocabulary.UserRole;
 import org.gbif.occurrence.annotation.mapper.CommentMapper;
 import org.gbif.occurrence.annotation.mapper.RuleMapper;
 import org.gbif.occurrence.annotation.model.Comment;
@@ -25,7 +24,6 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -38,6 +36,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+
+import static org.gbif.occurrence.annotation.controller.AuthAdvice.assertCreatorOrAdmin;
 
 @RestController
 @CrossOrigin(origins = "*")
@@ -83,9 +83,11 @@ public class RuleController implements Controller<Rule> {
 
   @Operation(summary = "Logical delete a rule")
   @DeleteMapping("/{id}")
-  @Secured("USER")
+  @Secured({"USER", "REGISTRY_ADMIN"})
   @Override
   public Rule delete(@PathVariable(value = "id") int id) {
+    Rule existing = ruleMapper.get(id);
+    assertCreatorOrAdmin(existing.getCreatedBy());
     ruleMapper.delete(id, getLoggedInUser());
     return ruleMapper.get(id);
   }
@@ -151,19 +153,8 @@ public class RuleController implements Controller<Rule> {
   @Secured({"USER", "REGISTRY_ADMIN"})
   public void deleteComment(@PathVariable(value = "commentId") int commentId) {
     Comment existing = commentMapper.get(commentId);
-    String username = getLoggedInUser();
-    if (!existing.getCreatedBy().equals(username)) {
-      boolean isAdmin =
-          SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
-              .anyMatch(r -> r.getAuthority().equals(UserRole.REGISTRY_ADMIN.toString()));
-
-      if (!isAdmin) {
-        throw new AuthAdvice.NotAuthorisedException(
-            "Only the comment author or a registry admin can delete a comment");
-      }
-      ;
-    }
-    commentMapper.delete(commentId, username);
+    assertCreatorOrAdmin(existing.getCreatedBy());
+    commentMapper.delete(commentId, getLoggedInUser());
   }
 
   @Operation(
